feat: add improvements
